Create new folder and index.qmd in your blog's posts directory

Create new folder and index.qmd in your blog's posts directory

Usage

newpost(
  headline,
  description = "",
  author = "",
  categories = "",
  root_dir = here::here(),
  the_date = Sys.Date()
)

Arguments

description

string with post description. Optional but recommended.

author

string with author name. Suggest including this in a code snippet, see vignette. OPtional.

categories

string in the format "category1, category2" as a single comma-separated string and NOT a vector.

root_dir

string your blog's root directory, defaults to here::here()

the_date

date or string in yyyy-mm-dd format

character

string with planned post headline. Required

Value

Creates a new folder in posts directory with yyy–mm-dd-headline-in-lowercase where yyyy-mm-dd is current system date